:recycle: refact: create utils file and obsever selected step method
