+++ controlsy/Sources/Core/Coordinators/Utils/ChildCoordinatorStep.swift
@@ -1,0 +1,13 @@
+//
+//  ChildCoordinatorStep.swift
+//  controlsy
+//
+//  Created by Gabriel Matheus on 16/11/22.
+//
+
+import Foundation
+
+enum Step {
+    case stop
+    case none
+}
