+++ controlsy/Sources/Core/Coordinators/Utils/NavigationType.swift
@@ -1,0 +1,14 @@
+//
+//  NavigationType.swift
+//  controlsy
+//
+//  Created by Gabriel Matheus on 16/11/22.
+//
+
+import Foundation
+
+enum NavigationType {
+    case onboarding
+    case session
+    case home
+}
